+++ items/items.js
@@ -1,11 +1,10 @@
-module.exports = {
-  repair,
-}
-
-function repair(item) {
-  if (item.durability !== 100) {
-    return {...item, durability: 100};
-  } else {
-    null;
+class Item {
+  constructor(name, type, durability = 100, enhancement = 0) {
+    this.name = name,
+      this.type = type.toUpperCase(),
+      this.durabilityValue = durability,
+      this.enhancementValue = enhancement;
   }
 }
+
+module.exports = Item;
